requirements.txt
# requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
python-dotenv
Flask-Login
Werkzeug
cryptography